requirements.txt
faiss-cpu==1.12.0
numpy==2.3.2
pandas==2.2.2
hdbscan==0.8.40 
light-embed==1.0.8